Move entrypoint.ts to providers
